Add static assertion for number of `enum dots_expansion_op`
