+++ src/dots.h
@@ -1,0 +1,32 @@
+#ifndef RLANG_INTERNAL_DOTS_H
+#define RLANG_INTERNAL_DOTS_H
+
+
+#define DOTS_CAPTURE_TYPE_MAX 3
+enum dots_capture_type {
+  DOTS_EXPR,
+  DOTS_QUO,
+  DOTS_VALUE
+};
+
+enum dots_expansion_op {
+  OP_EXPR_NONE,
+  OP_EXPR_UQ,
+  OP_EXPR_UQE,
+  OP_EXPR_UQS,
+  OP_EXPR_UQN,
+  OP_QUO_NONE,
+  OP_QUO_UQ,
+  OP_QUO_UQE,
+  OP_QUO_UQS,
+  OP_QUO_UQN,
+  OP_VALUE_NONE,
+  OP_VALUE_UQ,
+  OP_VALUE_UQE,
+  OP_VALUE_UQS,
+  OP_VALUE_UQN,
+  OP_DOTS_MAX
+};
+
+
+#endif
